Clean up code
Now using toolstrip item callbacks instead of using the item's name to determine what was clicked.
Button callbacks are now added from code instead of the events tab.
Known issue: When the application first starts it shows as a foreground app in task manager and can be seen when alt tabbing, it goes into the background after opening and closing the settings.